+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/MainActivity.kt
@@ -1,10 +1,8 @@
 package jp.shiguredo.sora.sample.ui
 
 import android.Manifest
-import android.annotation.SuppressLint
 import android.annotation.TargetApi
 import android.content.Intent
-import android.os.Build
 import android.os.Bundle
 import android.support.design.widget.Snackbar
 import android.support.v7.app.AlertDialog